mod: splits typechecks methods in another module.
